feat: add occupations dataset builder, docs
