+++ data/esco/README.md
@@ -1,0 +1,21 @@
+# Esco server quick start guide
+
+## Prerequisites
+
+-   nodejs
+-   java runtime
+
+## Install
+
+-   run `npm install`
+-   Download the esco v1.1.0 server from https://esco.ec.europa.eu/en/use-esco/download
+-   Unzip the downloaded file to `./server` folder
+
+## Usage
+
+-   start server with `npm run serve` (~30s boot time)
+-   run data building scripts with `npm run build:<feat>` eg. `npm run build:skills`
+
+## Documentation
+
+Navigate to http://localhost:8080/doc/esco_api_doc.html
